reworked some things

+++ src/GameController/Block.java
@@ -3,13 +3,13 @@ package GameController;
 import javafx.scene.paint.Color;
 
 public class Block {
-    public double Y_POSITION;
-    public double X_POSITION;
-    public double WIDTH;
-    public double HEIGHT;
+    public int Y_POSITION;
+    public int X_POSITION;
+    public int WIDTH;
+    public int HEIGHT;
     public Color COLOR;
 
-    public Block(double y, double x, double width, double height, Color c){
+    public Block(int y, int x, int width, int height, Color c){
         this.Y_POSITION = y;
         this.X_POSITION = x;
         this.WIDTH = width;
